refactor: split 3 endpoint files (1000-1200 LOC each) into packages
Split 3 top-level endpoint files using package+router pattern:

| File | Before | After (shim) | Modules | Endpoints |
|---|---|---|---|---|
| admin_departments.py | 1117 | 10 | 3 (admin_departments/) | 18 |
| user_management.py | 1103 | 10 | 4 (user_management/) | 21 |
| services.py | 1010 | 10 | 3 (services_ep/) | 17 |
| **Total** | **3230** | **30** | **10** | **56** |

All public APIs preserved. Ruff: 0 errors. 34 tests pass.
API router: 1128 routes, 0 duplicates.
